// src/app/models/masters-mpc/measurementunits.ts
export class measurementunits {
    id: number = 0;
    name: string = '';
    abbreviation: string = '';
    idGroupingUnitofMeasure: number = 0;
    groupingUnitofMeasure: string = '';
    createdByUser: string = '';
    createdByUserId: number = 0;
    updatedByUser: string = '';
    active: boolean = false;
    activeGroupingUnitofMeasure: boolean;
    initialSetup: boolean = false;
}